pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xclaw-ag-memory-guard"
version = "1.0.0"
description = "Memory protection for OpenClaw agents - validates memory operations for security and integrity"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "XClaw AgentGuard Security Team"}
]
keywords = [
    "openclaw",
    "security",
    "memory-protection",
    "data-integrity"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
requires-python = ">=3.12"
dependencies = [
    "xclaw-agentguard>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/neil-njcn/xclaw-ag-memory-guard"
Repository = "https://github.com/neil-njcn/xclaw-ag-memory-guard.git"
Issues = "https://github.com/neil-njcn/xclaw-ag-memory-guard/issues"

[project.entry-points."openclaw.skills"]
memory_guard = "xclaw_ag_memory_guard:MemoryGuardSkill"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
